Finished Merging
Fixed symmetry option for radial.
Symmetry can now be used for straight bars.

Bars are now generated in a Bizualizer Collection, and parented to an empty. This empty can be moved scaled and rotated, and if the bars are regenerated they get reassigned to the empty.
Multiple sets of bars can be made just the Bar Name needs to be changed and a new empty is made.